[GPU] Rename RMS FP16 pass files to all-rms naming
### Details:

- Renamed disable_fp16_comp_for_rms_norm_block.* to disable_fp16_comp_for_all_rms.* to match pass scope (all RMS nodes).

- Updated includes in transformations_pipeline.cpp and disable_fp16_compression_rms_test.cpp.

- No behavior change in pass logic; this is naming and discoverability cleanup requested by code review.

### Tickets:

- *CVS-187992*
