+++ src/plugins/intel_gpu/src/plugin/transformations/disable_fp16_comp_for_all_rms.hpp
